# fix: cleanup some unnecessary dependencies