--- a/Cargo.toml
+++ b/Cargo.toml
@@ -62,13 +62,11 @@ harness = false
 test = true
 
 [dependencies]
-arrayvec = { version = "0.7.6", default-features = false }
 bitvec.workspace = true
 bytes = { version = "1.7.2", default-features = false }
 chrono.workspace = true
 either = { version = "1.13.0", default-features = false }
 hashbrown = "0.14.5"
-konst = { version = "0.3.9", default-features = false }
 nom = { version = "7.1.3", default-features = false, features = ["alloc"] }
 nom-bitvec = { package = "bitvec-nom2", version = "0.2.1" }
 num-bigint = { version = "0.4.6", default-features = false }