--- a/Cargo.toml
+++ b/Cargo.toml
@@ -28,8 +28,6 @@ chrono = { version = "0.4.38", default-features = false, features = ["alloc"] }
 bitvec = { version = "1.0.1", default-features = false, features = ["alloc"] }
 
 [features]
-default = ["macros"]
-macros = ["rasn-derive"]
 std = []
 backtraces = ["std", "snafu/backtrace"]
 compiler = ["rasn-compiler"]
@@ -79,7 +77,7 @@ once_cell = { version = "1.20.2", default-features = false, features = [
     "alloc",
 ] }
 rasn-compiler = { version = "0.5.3", optional = true }
-rasn-derive = { version = "0.20", path = "macros", optional = true }
+rasn-derive = { version = "0.20", path = "macros" }
 snafu = { version = "0.8.5", default-features = false, features = [
     "rust_1_81",
 ] }