Fixes #359 

Seems like `rayon` is only used for processing constraints in parallel and typically the amount of constraints is rather small.
I could not measure any drastic performance difference when compiling whole workspace, so I just removed `rayon`:

```
# With rayon
cargo build --workspace --all-targets  142.49s user 17.37s system 742% cpu 21.517 total

# Without rayon
cargo build --workspace --all-targets  141.91s user 17.11s system 752% cpu 21.136 total
```


Also `arrayvec` has been removed as I could not find any usage for it.

There is constant equivalent for current `konst` dependency usage, so this removes that as well.