@Nicceboy 2024-10-29
I also removed the `macros` feature since it is not currently possible to compile `rasn` without it. Or should we rework `Open` type before doing so?